requirements.txt
click==6.6
nltk==3.2.1
oauthlib==2.0.0
pkg-resources==0.0.0
python-dotenv==0.6.0
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
textblob==0.11.1
tweepy==3.5.0
